Refactor vcf tests into separate file
